--- example/BuildSuccess16/test/example/FooTest.java
@@ -1,15 +1,0 @@
-package example;
-
-import static org.junit.Assert.*;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import org.junit.Test;
-
-public class FooTest{
-  @Test
-  public void test01() {
-    final Foo foo = new Foo();
-    final Path path = Paths.get("./resources/test.txt");
-    assertEquals(foo.readFile(path), "Hello World");
-  }
-}
