+++ example/BuildSuccess17/src/example/Foo.java
@@ -3,15 +3,17 @@ package example;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.List;
 
 public class Foo {
 
-  public void write(final Path path, final List<String> list) {
+  public String readFile(final Path path) {
+    String result = "";
     try {
-      Files.write(path, list);
+      result = Files.readAllLines(path).get(0);
     } catch (final IOException e) {
       e.printStackTrace();
     }
+
+    return result;
   }
 }
